refactor: Move test utiliy functions
